python/agent-hub/mem0-search/pyproject.toml
[tool.poetry]
name = "mem0-search"
version = "0.1.1"
authors = [
    "ZongHuan Wu ",
    "Cheng Chen",
]
description = "Use Mem0 to retrieve memory"
license = "MIT License"
homepage = "https://github.com/moxin-org/mofa"
documentation = "https://github.com/mem0ai/mem0"
readme = "README.md"
packages = [{ include = "mem0_search" }]

[tool.poetry.dependencies]
pyarrow = ">= 5.0.0"
mem0ai = ">=0.1.97"
flask = "*"
[tool.poetry.scripts]
mem0-search = "mem0_search.main:main"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
